final changes. Add asm files for ap.c and example.c. Add generated executable

+++ ap.c
@@ -2,11 +2,11 @@
 #include <errno.h>
 #include <stdatomic.h>
 #include <stddef.h>
+#include <stdint.h>
 
 typedef struct {
- // might want to store the user-provided address and the trampoline address
   uintptr_t addr;
-  void (*handler)(uint8_t);
+  void (*handler)(uintptr_t);
 } ap_entry_t;
 
 typedef struct {
@@ -16,14 +16,15 @@ typedef struct {
 
 ap_entry_t ap_table[AP_MAX_ENTRIES];
 uint64_t free_mask = ~0ULL; //1 means free, 0 means used
-int acquire_ap_idx(uintptr_t ap_addr, void(*ap_handler)(uint8_t));
+int acquire_ap_idx(uintptr_t ap_addr, void(*ap_handler)(uintptr_t));
 int alloc_slot();
 void dealloc_slot(int index);
 ap_flags_t parse_ap_flags(uint8_t ap_flags);
 void configure_ap(int index, uint8_t bool_active, uint8_t bool_oneshot,void *target_addr, void *trigger_addr);
 void ap_set_active(int index, uint8_t active);
 
-int ap_reg(uintptr_t ap_addr, void(*ap_handler) (uint8_t), uint8_t ap_flags) {
+
+int ap_reg(uintptr_t ap_addr, void(*ap_handler) (uintptr_t), uint8_t ap_flags) {
   if (ap_handler == NULL || ap_addr == 0) {
     errno = EINVAL; 
     return -1;
@@ -45,7 +46,8 @@ int ap_reg(uintptr_t ap_addr, void(*ap_handler) (uint8_t), uint8_t ap_flags) {
   return 0;
 }
 
-void ap_ureg(uintptr_t ap_addr, void(*ap_handler)(uint8_t)) {
+void ap_ureg(uintptr_t ap_addr, void(*ap_handler)(uintptr_t)) {
+  //could use a hash map here, but chose to do linear search instead since the number of entries is so small
   for (int i = 0; i < AP_MAX_ENTRIES; i++) {
     if (ap_table[i].addr == ap_addr && ap_table[i].handler == ap_handler) {
       ap_table[i].addr = 0;
@@ -58,7 +60,8 @@ void ap_ureg(uintptr_t ap_addr, void(*ap_handler)(uint8_t)) {
 }
 
 
-int acquire_ap_idx(uintptr_t ap_addr, void(*ap_handler)(uint8_t)) {
+//returns the index of the entry on success and -1 on failure.
+int acquire_ap_idx(uintptr_t ap_addr, void(*ap_handler)(uintptr_t)) {
     for (int i = 0; i < AP_MAX_ENTRIES; i++) {
         if (ap_table[i].addr == ap_addr && ap_table[i].handler == ap_handler) {
           return i;
@@ -79,7 +82,7 @@ int acquire_ap_idx(uintptr_t ap_addr, void(*ap_handler)(uint8_t)) {
 }
 
 int alloc_slot() {
-    int slot = __builtin_ctzll(free_mask);  // single instruction
+    int slot = __builtin_ctzll(free_mask);
     free_mask &= ~(1ULL << slot);
     return slot;
 }
